requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
pymupdf
beautifulsoup4
arxiv
wikipedia
chromadb
faiss-cpu
langchain_huggingface
sentence_transformers
langchain_openai
langchain_ollama
streamlit
langchain_groq
langserve
langchain_chroma
validators
youtube_transcript_api
pytube
unstructured